refactor: Modularize WaterUseAgent and enhance Canal model
This commit refactors the canal diversion simulation example based on user feedback to improve modularity and reusability.

Key changes:
- The `WaterUseAgent` class has been moved from the example script into its own reusable module at `swp/simulation_identification/disturbances/water_use_agent.py`.
- The `Canal` model in `swp/simulation_identification/physical_objects/canal.py` has been enhanced to subscribe to a message bus topic for data-driven inflows, enabling more flexible disturbance modeling.
- The main example script, `swp/examples/canal_diversion_simulation.py`, has been updated to import the `WaterUseAgent` from its new location and to use the new data-driven inflow capability of the `Canal` model.

This addresses the architectural feedback to define agents as reusable base classes and not within example code.
